Fix(File Saving): Prevent existing additional files from being overwritten.
When saving additional files like cue sheets or logs, Picard would overwrite files with the same name in the destination directory. This change ensures that existing files are not overwritten by adding a numerical suffix to the new file if a name collision occurs.

Fixes PICARD-271.

+++ picard/file.py
@@ -657,6 +657,9 @@ class File(MetadataItem):
             if self.tagger.files.get(decode_filename(old_file_path)):
                 log.debug("File loaded in the tagger, not moving %r", old_file_path)
                 continue
+            if os.path.exists(new_file_path):
+                log.warning("File %r already exists, not moving %r", new_file_path, old_file_path)
+                continue
             log.debug("Moving %r to %r", old_file_path, new_file_path)
             try:
                 shutil.move(old_file_path, new_file_path)
